Stop processing after match
At the moment, unlike procmail, the rules are all executed sequentially
regardless of flags etc..

There is an edge case where an item might be matched by more than one
rule.  In this circumstance, the last rule in the list "wins".

It logically makes more sense to me for the first match to "win" -- as
in the example docs currently, I have a DELETE rule for cancellations as
the first rule; I could move this to last, but top down ordering makes
more sense to me.

There are also runtime benefits as you only need to keep attempting
matches until the first one hits (as opposed to processing every rule
each time).